+++ src/providers/mixpanel/mixpanel.js
@@ -56,11 +56,11 @@ analytics.addProvider('Mixpanel', {
         // Alias the traits' keys with dollar signs for Mixpanel's API.
         if (traits) {
             analytics.utils.alias(traits, {
-                'email'     : '$email',
-                'name'      : '$name',
-                'username'  : '$username',
-                'lastSeen'  : '$lastSeen',
-                'createdAt' : '$created'
+                'email'    : '$email',
+                'name'     : '$name',
+                'username' : '$username',
+                'lastSeen' : '$lastSeen',
+                'created'  : '$created'
             });
         }